Started working on many to one and composite keys

+++ src/main/java/no/ntnu/idatt2106/repository/CategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.CategoryDAO;
 
-public interface CategoryRepository extends JpaRepository<CategoryDAO, Integer> {
+public interface CategoryRepository extends JpaRepository<CategoryDAO, Long> {
     
 }
