+++ src/main/java/no/ntnu/idatt2106/repository/ChatMessageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.ChatMessageDAO;
 
-public interface ChatMessageRepository extends JpaRepository<ChatMessageDAO, Integer> {
+public interface ChatMessageRepository extends JpaRepository<ChatMessageDAO, Long> {
     
 }
