+++ src/main/java/no/ntnu/idatt2106/repository/CommentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.CommentDAO;
 
-public interface CommentRepository extends JpaRepository<CommentDAO, Integer> {
+public interface CommentRepository extends JpaRepository<CommentDAO, Long> {
     
 }
