+++ src/main/java/no/ntnu/idatt2106/repository/FeedEntryPictureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.FeedEntryPictureDAO;
 
-public interface FeedEntryPictureRepository extends JpaRepository<FeedEntryPictureDAO, Integer> {
+public interface FeedEntryPictureRepository extends JpaRepository<FeedEntryPictureDAO, Long> {
     
 }
