+++ src/main/java/no/ntnu/idatt2106/repository/FeedEntryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.FeedEntryDAO;
 
-public interface FeedEntryRepository extends JpaRepository<FeedEntryDAO, Integer> {
+public interface FeedEntryRepository extends JpaRepository<FeedEntryDAO, Long> {
     
 }
