+++ src/main/java/no/ntnu/idatt2106/repository/GroupRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.GroupDAO;
 
-public interface GroupRepository extends JpaRepository<GroupDAO, Integer> {
+public interface GroupRepository extends JpaRepository<GroupDAO, Long> {
     
 }
