+++ src/main/java/no/ntnu/idatt2106/repository/GroupRequestRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.GroupRequestDAO;
 
-public interface GroupRequestRepository extends JpaRepository<GroupRequestDAO, Integer> {
+public interface GroupRequestRepository extends JpaRepository<GroupRequestDAO, Long> {
     
 }
