+++ src/main/java/no/ntnu/idatt2106/repository/ListingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.ListingDAO;
 
-public interface ListingRepository extends JpaRepository<ListingDAO, Integer> {
+public interface ListingRepository extends JpaRepository<ListingDAO, Long> {
     
 }
