+++ src/main/java/no/ntnu/idatt2106/repository/RatingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.RatingDAO;
 
-public interface RatingRepository extends JpaRepository<RatingDAO, Integer> {
+public interface RatingRepository extends JpaRepository<RatingDAO, Long> {
     
 }
