+++ src/main/java/no/ntnu/idatt2106/repository/UserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.UserDAO;
 
-public interface UserRepository extends JpaRepository<UserDAO, Integer> {
+public interface UserRepository extends JpaRepository<UserDAO, Long> {
     
 }
